add playlist detail page
module: app

on branch: master

+++ app/src/main/java/com/jascal/galatea/net/service/MusicService.kt
@@ -19,5 +19,5 @@ interface MusicService {
      * */
 
     @GET("playlist/detail")
-    fun getPlaylistDetail(@Query("id") playlistID: Int): Observable<PlaylistDetailResponse>
+    fun getPlaylistDetail(@Query("id") playlistID: Long): Observable<PlaylistDetailResponse>
 }
